server/requirements.txt
alembic==1.11.0
fastapi==0.95.2
psycopg2==2.9.6
pydantic==1.10.7
python-dotenv==1.0.0
sqlalchemy==2.0.13
typer[all]==0.9.0
uvicorn==0.22.0
